saveme/model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
matplotlib>=3.7.2
joblib>=1.3.2
requests>=2.31.0
